@@ -9,10 +9,15 @@
     "googleapis": "^18.0.0",
     "jquery": "^3.2.1",
     "material-ui": "^0.17.1",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
+    "prop-types": "^15.5.9",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
+    "react-redux": "^5.0.4",
+    "react-router": "^4.1.1",
+    "react-router-dom": "^4.1.1",
     "react-tap-event-plugin": "^2.0.1",
-    "react-vendor-prefix": "^0.0.1"
+    "react-vendor-prefix": "^0.0.1",
+    "redux": "^3.6.0"
   },
   "devDependencies": {
     "flow-bin": "^0.42.0",
