@@ -1,4 +1,5 @@
 import React, {Component} from 'react';
+import PropTypes from 'prop-types';
 import VendorPrefix from 'react-vendor-prefix';
 import assign from 'object-assign';
 
@@ -33,9 +34,9 @@ export default class Layout extends Component {
 }
 
 Layout.propTypes = {
-  type: React.PropTypes.string.isRequired,
-  className: React.PropTypes.string,
-  style: React.PropTypes.object
+  type: PropTypes.string.isRequired,
+  className: PropTypes.string,
+  style: PropTypes.object
 };
 
 Layout.defaultProps = {
